Add lock label to GitHub pull requests after plan
Requires a lock label to already be created on the repository. Only
implemented for GitHub. Other VCS hosts just stubbed out in the Client
interface.

Adds a --label-lock-pr flag to the atlantis server command, but the flag
is purely cosmetic in this commit. Does not actually influence any
meaningful logic.